```text Add functions for creating posts and uploading images
This commit adds two new functions to the upload.py file. The `create_post` function is used to create a new post on a WordPress site, while the `upload_image` function is used to upload an image to the site. These functions are necessary for publishing content with text and images.